moved out detectedAV function
