+++ packages/web-app/src/Store.ts
@@ -122,13 +122,13 @@ export class RootStore {
     this.vault = new VaultStore(axios, this.balance, this.rewards)
     this.version = new VersionStore(this, axios)
     this.engagement = new EngagementStore(this, axios)
-    this.zendesk = new Zendesk(axios, featureManager, this.auth, this.native, this.analytics)
+    this.zendesk = new Zendesk(axios, featureManager, this.auth, this.analytics)
     this.storefront = new StorefrontStore(axios)
     this.bonuses = new BonusStore(this, axios)
     this.seasons = new SeasonsStore(axios)
     this.onboarding = new OnboardingStore(this)
     this.onboardingAutoStart = new OnboardingAutoStartStore(this)
-    this.onboardingAntivirus = new OnboardingAntivirusStore(this)
+    this.onboardingAntivirus = new OnboardingAntivirusStore(this, this.native)
     this.startButtonUI = new StartButtonUIStore(this)
     this.machineSettingsUI = new MachineSettingsUIStore(this)
     this.detectedHardwareUIStore = new DetectedHardwareUIStore(this)
